requirements.txt
openai-whisper
streamlit
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dateutil
dateparser
python-dotenv
ffmpeg
